requirements.txt
-e git+https://github.com/mindee/doctr.git#egg=python-doctr[torch]
openai==0.28
python-dotenv
fastapi
uvicorn
apscheduler
pymongo